WIP: Add new build.sh for all apple platforms
This adds a new build.sh script which works for iOS, macOS and tvOS
and can be used by VLCKit build script instead of having to re-implement
basically all of build.sh itself.

This is still a work in progress, some things are lacking still,
like setting the right env variables or adapting contrib bootstrap
to behave properly for iOS or tvOS.
Additionally builds with static modules or multiple archs are not yet
supported, but especially the latter should be trivial to add.